Change "generate_entities" mix task to use another file
This is because we now have some rules in the main file.
