+++ lib/mix/tasks/generate_entities.ex
@@ -2,7 +2,7 @@ defmodule Mix.Tasks.GenerateEntities do
   @shortdoc "Generate the entities module"
 
   @json_entities_path "priv/entities.json"
-  @destination_module_path "lib/floki/entities.ex"
+  @destination_module_path "lib/floki/entities/codepoints.ex"
 
   use Mix.Task
 
@@ -14,7 +14,7 @@ defmodule Mix.Tasks.GenerateEntities do
     {:ok, json} = Jason.decode(content)
 
     headers = """
-    defmodule Floki.Entities do
+    defmodule Floki.Entities.Codepoints do
     # This file was generated by "Mix.Tasks.GenerateEntities"
 
     @moduledoc false
@@ -52,6 +52,6 @@ defmodule Mix.Tasks.GenerateEntities do
 
     File.write!(@destination_module_path, contents)
 
-    Mix.shell().info("Entities module is located in lib/floki/entities.ex")
+    Mix.shell().info("Entities module is located in lib/floki/entities/codepoints.ex")
   end
 end
